Updated all changes
added Comments
Auto mapper
Unit test
Proper naming convention
Database project added
Table data insertion script added
