+++ TestingAndCalibrationLabs/TestingAndCalibrationLabs.DB/Scripts/InserSampleData.sql
@@ -1,0 +1,3 @@
+INSERT INTO SampleTable(Name,Description,IsDeleted) values ('EyeBrow','EyeBrow','0')
+INSERT INTO SampleTable(Name,Description,IsDeleted) values ('Eye','Eye','0')
+INSERT INTO SampleTable(Name,Description,IsDeleted) values ('india','aaa','0')
