+++ TestingAndCalibrationLabs/TestingAndCalibrationLabs.sln
@@ -11,7 +11,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Database", "Database", "{51
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{7EE1F4A1-D042-4AF5-A84F-9BD7A5B97EAD}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestingAndCalibrationLabs.Tests", "TestingAndCalibrationLabs\TestingAndCalibrationLabs.Tests.csproj", "{5782CEC0-A093-41DA-B015-09C1993841F0}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestingAndCalibrationLabs.Tests", "TestingAndCalibrationLabs\TestingAndCalibrationLabs.Tests.csproj", "{5782CEC0-A093-41DA-B015-09C1993841F0}"
+EndProject
+Project("{00D1A9C2-B5F0-4AF3-8072-F6C62B433612}") = "TestingAndCalibrationLabs.DB", "TestingAndCalibrationLabs.DB\TestingAndCalibrationLabs.DB.sqlproj", "{7531576F-6091-474B-BC66-F1AC77D593B4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -31,12 +33,19 @@ Global
 		{5782CEC0-A093-41DA-B015-09C1993841F0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5782CEC0-A093-41DA-B015-09C1993841F0}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5782CEC0-A093-41DA-B015-09C1993841F0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7531576F-6091-474B-BC66-F1AC77D593B4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7531576F-6091-474B-BC66-F1AC77D593B4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7531576F-6091-474B-BC66-F1AC77D593B4}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
+		{7531576F-6091-474B-BC66-F1AC77D593B4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7531576F-6091-474B-BC66-F1AC77D593B4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7531576F-6091-474B-BC66-F1AC77D593B4}.Release|Any CPU.Deploy.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{5782CEC0-A093-41DA-B015-09C1993841F0} = {7EE1F4A1-D042-4AF5-A84F-9BD7A5B97EAD}
+		{7531576F-6091-474B-BC66-F1AC77D593B4} = {51F24674-1B4D-4BCE-BE5D-953DDAB78CE5}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {59F2B9C4-C240-453F-97B0-9B5DB91EE197}
